Also manage annotations ids
