+++ .env.ci
@@ -7,6 +7,7 @@ DEMARCHES_SIMPLIFIEES_ID=1
 # URL for the candidates to start their demarche. (Do not use quotes around the URL on Scalingo)
 DEMARCHES_SIMPLIFIEES_URL=https://demarches-simplifiees.fr/
 DEMARCHES_SIMPLIFIEES_CHAMPS=[["Q2hhbXAtMTY1OTUxMw==", "departement"], ["Q2hhbXAtMTYzMDQxNw==", "diploma"], ["Q2hhbXAtMTYzOTE2OQ==", "diplomaYear"], ["Q2hhbXAtMTYyNjk4Nw==", "adeli"], ["Q2hhbXAtMTYzOTE2Nw==", "training"], ["Q2hhbXAtMTYwMzgwNQ==", "phone"], ["Q2hhbXAtMTYyNzkzOQ==", "address"], ["Q2hhbXAtMTYwMTE4Ng==", "email"], ["Q2hhbXAtMTYzOTUyNA==", "teleconsultation"], ["Q2hhbXAtMTY2MDM0Nw==", "languages"], ["Q2hhbXAtMTYzOTQwMQ==", "website"], ["Q2hhbXAtMTYzMDU3Nw==", "description"]]
+DEMARCHES_SIMPLIFIEES_ANNOTATIONS=[["Q2hhbXAtMTY1NzQwMw==", "verifiee"]]
 
 # To set in scalingo for production - is set in docker-compose files for dev
 SECRET='production_value_should_be_set_in_.env'
