+++ .env.sample
@@ -10,6 +10,7 @@ DEMARCHES_SIMPLIFIEES_AUTO_ACCEPT_DEPARTMENTS=['03 - Allier', '07 - Ardèche', '
 DEMARCHES_SIMPLIFIEES_INSTRUCTOR=""
 DEMARCHES_SIMPLIFIEES_ANNOTATION_VERIFIEE=""
 DEMARCHES_SIMPLIFIEES_CHAMPS=[["Q2hhbXAtMTY1OTUxMw==", "departement"], ["Q2hhbXAtMTYzMDQxNw==", "diploma"], ["Q2hhbXAtMTYzOTE2OQ==", "diplomaYear"], ["Q2hhbXAtMTYyNjk4Nw==", "adeli"], ["Q2hhbXAtMTYzOTE2Nw==", "training"], ["Q2hhbXAtMTYwMzgwNQ==", "phone"], ["Q2hhbXAtMTYyNzkzOQ==", "address"], ["Q2hhbXAtMTYwMTE4Ng==", "email"], ["Q2hhbXAtMTYzOTUyNA==", "teleconsultation"], ["Q2hhbXAtMTY2MDM0Nw==", "languages"], ["Q2hhbXAtMTYzOTQwMQ==", "website"], ["Q2hhbXAtMTYzMDU3Nw==", "description"]]
+DEMARCHES_SIMPLIFIEES_ANNOTATIONS=[["Q2hhbXAtMTY1NzQwMw==", "verifiee"]]
 
 # To set in scalingo for production - is set in docker-compose files for dev
 # SECRET='cat'
